+++ src/moonlight_steam_sync/__init__.py
@@ -2,7 +2,7 @@
 
 from importlib import metadata
 
-__version__ = "0.3.1"
+__version__ = "0.4.0"
 
 
 def version() -> str: